bin/CONCOCT-1.1.0/requirements.txt
argparse>=1.2.1
bcbio-gff>=0.4,<=0.6.4; python_version < '3.0'
bcbio-gff>=0.4; python_version >= '3.0'
biopython>=1.63,<=1.72; python_version < '3.0'
biopython>=1.63; python_version >= '3.0'
nose>=1.3.0
numpy>=1.8.0
pandas>=0.13.0
python-dateutil>=2.2
pytz>=2013.9
scikit-learn>=0.14.1
scipy>=0.13.3
cython>=0.28.5
